+++ R/cansim_sql.R
@@ -42,6 +42,8 @@ get_cansim_sqlite <- function(cansimTableNumber, language="english", refresh=FAL
     else
       message(paste0("Acc",intToUtf8(0x00E9),"der au produit ", cleaned_number, " CANSIM NDM de Statistique Canada"))
     url=paste0("https://www150.statcan.gc.ca/n1/tbl/csv/",file_path_for_table_language(cansimTableNumber,language),".zip")
+
+    time_check <- Sys.time()
     response <- get_with_timeout_retry(url,path=path,timeout=timeout)
     if (is.null(response)) return(response)
     data <- NA
@@ -51,6 +53,7 @@ get_cansim_sqlite <- function(cansimTableNumber, language="english", refresh=FAL
     if (is.null(uzp)) uzp <- "internal"
     utils::unzip(path,exdir=exdir,unzip=uzp)
     unlink(path)
+
     if(cleaned_language=="eng") {
       message("Parsing data")
       delim <- ","
@@ -60,6 +63,7 @@ get_cansim_sqlite <- function(cansimTableNumber, language="english", refresh=FAL
       delim <- ";"
       value_column="VALEUR"
     }
+
 
     meta <- suppressWarnings(readr::read_delim(file.path(exdir, paste0(base_table, "_MetaData.csv")),
                                                delim=delim,
@@ -81,9 +85,6 @@ get_cansim_sqlite <- function(cansimTableNumber, language="english", refresh=FAL
 
     to_drop <- intersect(headers,"TERMINATED") # not in use yet
 
-    meta1 <- readRDS(paste0(meta_base_path,"1"))
-    meta2 <- readRDS(paste0(meta_base_path,"2"))
-    meta3 <- readRDS(paste0(meta_base_path,"3"))
 
     scale_string <- ifelse(language=="fr","IDENTIFICATEUR SCALAIRE","SCALAR_ID")
     value_string <- ifelse(language=="fr","VALEUR","VALUE")
@@ -92,20 +93,15 @@ get_cansim_sqlite <- function(cansimTableNumber, language="english", refresh=FAL
     dimension_name_column <- ifelse(cleaned_language=="eng","Dimension name","Nom de la dimension")
     geography_column <- ifelse(cleaned_language=="eng","Geography",paste0("G",intToUtf8(0x00E9),"ographie"))
     data_geography_column <- ifelse(cleaned_language=="eng","GEO",paste0("G",intToUtf8(0x00C9),"O"))
+    coordinate_column <- ifelse(cleaned_language=="eng","COORDINATE",paste0("COORDONN",intToUtf8(0x00C9),"ES"))
 
-    geo_column <- intersect(pull(meta2,dimension_name_column),geography_column)
+    meta2 <- readRDS(paste0(meta_base_path,"2"))
+    geo_column_pos <- which(pull(meta2,dimension_name_column)==geography_column)
 
-    if (length(geo_column)==1) {
-      classification_code_column <- ifelse(cleaned_language=="eng","Classification Code","Code sur la classification")
-      hierarchy_column <- ifelse(cleaned_language=="eng","Hierarchy",paste0("Hi",intToUtf8(0x00E9),"rarchie"))
-
-      meta_x <- readRDS(paste0(meta_base_path,"_column_",geo_column))
-      classification_lookup <- rlang::set_names(meta_x[[classification_code_column]],meta_x$name)
-      classification_lookup_id <- rlang::set_names(meta_x[[classification_code_column]],meta_x$`Member ID`)
-      hierarchy_lookup <- rlang::set_names(meta_x[[hierarchy_column]],meta_x$name)
-      hierarchy_lookup_id <- rlang::set_names(meta_x[[hierarchy_column]],meta_x$`Member ID`)
+    if (length(geo_column_pos)==1) {
+      hierarchy_prefix <- ifelse(cleaned_language=="eng","Hierarchy for",paste0("Hi",intToUtf8(0x00E9),"rarchie pour"))
+      hierarchy_name <- paste0(hierarchy_prefix," ", data_geography_column)
     }
-
 
     csv2sqlite(file.path(exdir, paste0(base_table, ".csv")),
                sqlite_file = sqlite_path,
@@ -115,22 +111,23 @@ get_cansim_sqlite <- function(cansimTableNumber, language="english", refresh=FAL
                delim = delim,
                transform=function(data){
                  data <- data %>%
-                   dplyr::mutate_at(value_string,as.numeric) %>%
-                   dplyr::mutate(val_norm=!!as.name(value_string)*(`^`(10,as.integer(!!as.name(scale_string)))))
-                 if (length(geo_column)==1)
+                   dplyr::mutate_at(value_string,as.numeric)
+                 if (length(geo_column_pos)==1)
                    data <- data %>%
-                     dplyr::mutate(GeoUID=gsub("\\[|\\]","",as.character(classification_lookup[.data[[data_geography_column]]])))
+                     fold_in_metadata_for_columns(meta_base_path,geography_column) %>%
+                     select(-!!as.name(hierarchy_name))
                  data
                })
 
     unlink(exdir,recursive = TRUE)
 
+    date_field=ifelse(cleaned_language=="fra",paste0("P",intToUtf8(0x00C9),"RIODE DE R",intToUtf8(0x00C9),"F",intToUtf8(0x00C9),"RENCE"),"REF_DATE")
 
-    fields <- meta2$`Dimension name` %>%
-      gsub("^Geography$","GEO",.) %>%
-      c(.,"REF_DATE","DGUID")
+    fields <- pull(meta2,dimension_name_column) %>%
+      gsub(geography_column,data_geography_column,.) %>%
+      c(.,date_field,"DGUID")
 
-    if (length(geo_column)==1) fields <- c(fields,"GeoUID")
+    if (length(geo_column_pos)==1) fields <- c(fields,"GeoUID")
 
     con <- DBI::dbConnect(RSQLite::SQLite(), dbname=sqlite_path)
     for (field in fields) {
@@ -140,7 +137,8 @@ get_cansim_sqlite <- function(cansimTableNumber, language="english", refresh=FAL
     DBI::dbDisconnect(con)
 
     # saving timestamp
-    saveRDS(strftime(Sys.time(),format=TIME_FORMAT),paste0(meta_base_path,"_time"))
+    saveRDS(strftime(time_check,format=TIME_FORMAT),paste0(meta_base_path,"_time"))
+
 
   } else {
     if (cleaned_language=="eng")
@@ -149,7 +147,7 @@ get_cansim_sqlite <- function(cansimTableNumber, language="english", refresh=FAL
       message(paste0("Lecture du produit ",cleaned_number)," de CANSIM NDM ",intToUtf8(0x00E0)," partir du cache.")
   }
 
-  if (have_custom_path) {
+  if (have_custom_path||TRUE) {
     meta_base_path <- paste0(base_path_for_table_language(cansimTableNumber,language,cache_path),".Rda")
     meta_grep_string <- basename(meta_base_path)
     meta_dir_name <- dirname(meta_base_path)
@@ -178,6 +176,48 @@ disconnect_cansim_sqlite <- function(connection){
   DBI::dbDisconnect(connection$src$con)
 }
 
+#' collect data from connection and normalize cansim table output
+#'
+#' @param connection connection to database
+#' @param replacement_value (Optional) the name of the column the manipulated value should be returned in. Defaults to adding the `val_norm` value field.
+#' @param normalize_percent (Optional) When \code{true} (the default) normalizes percentages by changing them to rates
+#' @param default_month The default month that should be used when creating Date objects for annual data (default set to "01")
+#' @param default_day The default day of the month that should be used when creating Date objects for monthly data (default set to "01")
+#' @param factors (Optional) Logical value indicating if dimensions should be converted to factors. (Default set to \code{false}).
+#' @param strip_classification_code (strip_classification_code) Logical value indicating if classification code should be stripped from names. (Default set to \code{false}).
+#' @return a tibble with the normalized data
+#'
+#' @examples
+#' \donttest{
+#' library(dplyr)
+#'
+#' con <- get_cansim_sqlite("34-10-0013")
+#' data <- con %>%
+#'   filter(GEO=="Ontario") %>%
+#'   collect_and_normalize()
+#'
+#' disconnect_cansim_sqlite(con)
+#' }
+#' @export
+collect_and_normalize <- function(connection,
+                                  replacement_value="val_norm", normalize_percent=TRUE,
+                                  default_month="07", default_day="01",
+                                  factors=FALSE,strip_classification_code=FALSE){
+  c <- connection$ops
+  while ("x" %in% names(c)) {c <- c$x}
+  cansimTableNumber <- c[[1]] %>%
+    gsub("^cansim_|_fra$|_eng$","",.) %>%
+    cleaned_ndm_table_number()
+  connection %>%
+    collect() %>%
+    normalize_cansim_values(replacement_value=replacement_value,
+                            normalize_percent=normalize_percent,
+                            default_month=default_month,
+                            default_day=default_day,
+                            factors=TRUE,
+                            cansimTableNumber = cansimTableNumber)
+}
+
 
 #' List cached cansim SQLite database
 #'
@@ -194,8 +234,9 @@ list_cansim_sqlite_cached_tables <- function(cache_path=getOption("cansim.cache_
   result <- dplyr::tibble(path=dir(cache_path,"cansim_")) %>%
     dplyr::mutate(cansimTableNumber=gsub("^cansim_|_eng$|_fra$","",.data$path) %>% cleaned_ndm_table_number()) %>%
     dplyr::mutate(language=gsub("^cansim_\\d+_","",.data$path)) %>%
-    dplyr::mutate(timeCached=NA_character_) %>%
-    dplyr::select(.data$cansimTableNumber,.data$language,.data$timeCached,.data$path)
+    dplyr::mutate(timeCached=NA_character_,
+                  sqliteSize=NA_character_) %>%
+    dplyr::select(.data$cansimTableNumber,.data$language,.data$timeCached,.data$sqliteSize,.data$path)
 
   if (nrow(result)>0) {
     result$timeCached <- do.call("c",
@@ -204,6 +245,17 @@ list_cansim_sqlite_cached_tables <- function(cache_path=getOption("cansim.cache_
                                    if (length(pp)==1) {
                                      d<-readRDS(file.path(cache_path,p,pp))
                                      dd<- strptime(d,format=TIME_FORMAT)
+                                   } else {
+                                     dd <- NA
+                                   }
+                                   dd
+                                 }))
+    result$sqliteSize <- do.call("c",
+                                 lapply(result$path,function(p){
+                                   pp <- dir(file.path(cache_path,p),"\\.sqlite")
+                                   if (length(pp)==1) {
+                                     d<-file.size(file.path(cache_path,p,pp))
+                                     dd<- format_file_size(d,"auto")
                                    } else {
                                      dd <- NA
                                    }
@@ -252,6 +304,7 @@ remove_cansim_sqlite_cached_table <- function(cansimTableNumber,language=NULL,ca
 #' @param connection connection to database
 #' @param table_name sql table name
 #' @param field name of field to index
+#' @keywords internal
 create_index <- function(connection,table_name,field){
   field_index=paste0("index_",gsub("[^[:alnum:]]","_",field))
   query=paste0("CREATE INDEX IF NOT EXISTS ",field_index," ON ",table_name," (`",field,"`)")
@@ -277,6 +330,7 @@ create_index <- function(connection,table_name,field){
 #' @param na na character strings
 #' @param text_encoding encoding of csv file (default UTF8)
 #' @param delim (Optional) csv deliminator, default is ","
+#' @keywords internal
 csv2sqlite <- function(csv_file, sqlite_file, table_name, transform=NULL,chunk_size=5000000,
                        append=FALSE,col_types=NULL,na=c(NA,"..","","...","F"),
                        text_encoding="UTF8",delim = ",") {